requirements.txt
jupyter==1.0.0
mypy==1.7.0
pillow==10.1.0
click==8.1.7
pydantic==2.5.1
peewee==3.17.0
pytest==7.4.3
